+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Migrate unit test files to Typescript & Hardhat:
     - IexecPoco2
         - contribute (#108)
+    - IexecPoco/00_matchorders.js (#107)
     - Add `.test` suffix to unit test files (#106)
     - ENSIntegration (#105)
     - IexecOrderManagement (#101, #102, #103, #104)